Added python analysis
Added a Python analysis that takes into account each monitor's average photosensor to LSL delay. During analysis I noticed that the implementation of the experiment must be bugged because for all subjects, the majority of the trials were 40% target rate. I should've caught this earlier, but this is the kind of shenanigans that occur when facing intense deadlines, lessons to learn...